\name{gravmagsubs-package}
\alias{gravmagsubs-package}
\alias{gravmagsubs}
\docType{package}
\title{
  Gravitational and magnetic attraction of 3-D vertical rectangular prisms
}
\description{
  The package \code{gravmagsubs} provides tools for computing the gravitational and magnetic anomalies generated by 3-D vertical rectangular prisms at specific observation points.  The package consists of two functions:
  \itemize{
      \item \command{\link{rectprismgrav}} : Computes the gravitational attraction of 3-D right rectangular prisms.
      \item \command{\link{rectprismmag}} : Computes the magnetic effect of 3-D right rectangular prisms.
  }

  Each function can compute the total anomaly of a series of \code{N} prisms at \code{M} observation points.

  Each function also has a logical flag \code{bycell} (default \code{FALSE}).  If \code{bycell=TRUE}, the function returns the contribution from each individual prism.
}
\references{
\itemize{
\item Plouff, D., 1976, Gravity and magnetic fields of polygonal prisms and application to magnetic terrain corrections, Geophysics, v. 41, pp. 727--741, \doi{10.1190/1.1440645}.
}
}
